refactor(widget): move location action logic into a dedicated service
Take the location widget action feature out of widget.component.ts and
widget.models.ts so the widget component no longer carries it:

- new LocationService owns target resolution, alias fan-out, the attribute
  and time series writes, the browser geolocation flow and the
  startLiveLocation bridge payload; widget.component.ts keeps four
  delegating call sites
- new shared/models/location.models.ts holds the feature models: target
  entity config, location keys and the action descriptors
- resolve alias entities through the existing getAliasInfo and
  findEntityDataByQuery APIs rather than adding methods to
  entity.service.ts, alias-controller.ts and widget-api.models.ts, which
  are left untouched
- order alias entities by name so both the resolved page and the 100 entity
  fan-out cap are deterministic
