+++ ui-ngx/src/app/core/api/widget-api.models.ts
@@ -152,7 +152,6 @@ export interface IAliasController {
   getEntityAliasId(aliasName: string): string;
   getInstantAliasInfo(aliasId: string): AliasInfo;
   resolveSingleEntityInfo(aliasId: string): Observable<EntityInfo>;
-  resolveEntitiesInfo(aliasId: string): Observable<PageData<EntityInfo>>;
   resolveSingleEntityInfoForDeviceId(deviceId: string): Observable<EntityInfo>;
   resolveSingleEntityInfoForTargetDevice(targetDevice: TargetDevice): Observable<EntityInfo>;
   resolveDatasources(datasources: Array<Datasource>, singleEntity?: boolean, pageSize?: number): Observable<Array<Datasource>>;
