+++ ui-ngx/src/app/core/http/entity.service.ts
@@ -621,18 +621,6 @@ export class EntityService {
     );
   }
 
-  public findEntityInfosByEntityFilter(filter: EntityFilter, config?: RequestConfig): Observable<PageData<EntityInfo>> {
-    const query: EntityDataQuery = {
-      entityFilter: filter,
-      // Alias fan-out cap — savers report "saved X of Y" when the alias resolves more.
-      pageLink: createDefaultEntityDataPageLink(100),
-      entityFields: entityInfoFields
-    };
-    return this.findEntityDataByQuery(query, config).pipe(
-      map((data) => ({...data, data: data.data.map((entityData) => entityDataToEntityInfo(entityData))}))
-    );
-  }
-
   public getAliasFilterTypesByEntityTypes(entityTypes: Array<EntityType | AliasEntityType>): Array<AliasFilterType> {
     const authState = getCurrentAuthState(this.store);
     let allAliasFilterTypes: Array<AliasFilterType> = Object.values(AliasFilterType);
